--- src/interfaces/login.type.ts
@@ -1,5 +1,0 @@
-export type TLogin = { email: string; password: string };
-export interface ComparePasswordInterface {
-  password: string;
-  hashedPassword: string;
-}
