--- src/interfaces/request.interface.ts
@@ -1,7 +1,0 @@
-import { Request } from "express";
-import { UserAttribute } from "./user.type";
-
-export interface RequestWithUser extends Request {
-  user?: any;
-  query: any;
-}
