--- src/routes/masters.routes.ts
@@ -1,6 +1,0 @@
-// import * as express from "express";
-// import { getMasters } from "../controllers/getMasters.controller";
-// const router = express.Router();
-
-// router.get("/:app_code/:level1/:level2", getMasters);
-// export default router;
