+++ src/utils/passport.ts
@@ -1,35 +1,8 @@
-// import passport from "passport";
-// import passportJWT from "passport-jwt";
-
-// import constants from "../helpers/constants";
-// import User from "../models/user";
-// import { UserAttribute } from "../interfaces/user.type";
-
-// const JWTStrategy = passportJWT.Strategy;
-// const ExtractJWT = passportJWT.ExtractJwt;
-
-// passport.use(
-//   new JWTStrategy(
-//     {
-//       jwtFromRequest: ExtractJWT.fromAuthHeaderAsBearerToken(),
-//       secretOrKey: constants.AUTHENTICATION.APP_SECRET,
-//     },
-//     (jwtPayload, cb) =>
-//       User.findOne({ where: { email_id: jwtPayload.email_id } })
-//         .then((user: User | null) => {
-//           cb(null, user ? (user.dataValues as UserAttribute) : {});
-//         })
-//         .catch((err) => {
-//           cb(err);
-//         })
-//   )
-// );
-
 import passport from "passport";
 import passportJWT from "passport-jwt";
 import constants from "../helpers/constants";
 import User from "../models/user";
-import { UserAttribute } from "../interfaces/user.type";
+import { UserAttribute } from "../interfaces/user.interface";
 
 const JWTStrategy = passportJWT.Strategy;
 const ExtractJWT = passportJWT.ExtractJwt;
@@ -52,7 +25,6 @@ passport.use(
             cb(err);
           });
       } catch (err) {
-        // Handle errors gracefully
         return cb(err, false);
       }
     }
